Remove swagger 1.2 support
