+++ pyramid_swagger/api.py
@@ -7,7 +7,6 @@ from __future__ import absolute_import
 import copy
 import os.path
 
-import simplejson
 import yaml
 from bravado_core.spec import strip_xscope
 from six.moves.urllib.parse import urlparse
@@ -37,72 +36,6 @@ def register_api_doc_endpoints(config, endpoints, base_path='/api-docs'):
             endpoint.view,
             route_name=endpoint.route_name,
             renderer=endpoint.renderer)
-
-
-def build_swagger_12_endpoints(resource_listing, api_declarations):
-    """
-    :param resource_listing: JSON representing a Swagger 1.2 resource listing
-    :type resource_listing: dict
-    :param api_declarations: JSON representing Swagger 1.2 api declarations
-    :type api_declarations: dict
-    :rtype: iterable of :class:`pyramid_swagger.model.PyramidEndpoint`
-    """
-    yield build_swagger_12_resource_listing(resource_listing)
-
-    for name, filepath in api_declarations.items():
-        with open(filepath) as input_file:
-            yield build_swagger_12_api_declaration(
-                name, simplejson.load(input_file))
-
-
-def build_swagger_12_resource_listing(resource_listing):
-    """
-    :param resource_listing: JSON representing a Swagger 1.2 resource listing
-    :type resource_listing: dict
-    :rtype: :class:`pyramid_swagger.model.PyramidEndpoint`
-    """
-    def view_for_resource_listing(request):
-        # Thanks to the magic of closures, this means we gracefully return JSON
-        # without file IO at request time.
-        return resource_listing
-
-    return PyramidEndpoint(
-        path='',
-        route_name='pyramid_swagger.swagger12.api_docs',
-        view=view_for_resource_listing,
-        renderer='json')
-
-
-def build_swagger_12_api_declaration(resource_name, api_declaration):
-    """
-    :param resource_name: The `path` parameter from the resource listing for
-        this resource.
-    :type resource_name: string
-    :param api_declaration: JSON representing a Swagger 1.2 api declaration
-    :type api_declaration: dict
-    :rtype: :class:`pyramid_swagger.model.PyramidEndpoint`
-    """
-    # NOTE: This means our resource paths are currently constrained to be valid
-    # pyramid routes! (minus the leading /)
-    route_name = 'pyramid_swagger.swagger12.apidocs-{0}'.format(resource_name)
-    return PyramidEndpoint(
-        path='/{0}'.format(resource_name),
-        route_name=route_name,
-        view=build_swagger_12_api_declaration_view(api_declaration),
-        renderer='json')
-
-
-def build_swagger_12_api_declaration_view(api_declaration_json):
-    """Thanks to the magic of closures, this means we gracefully return JSON
-    without file IO at request time.
-    """
-    def view_for_api_declaration(request):
-        # Note that we rewrite basePath to always point at this server's root.
-        return dict(
-            api_declaration_json,
-            basePath=str(request.application_url),
-        )
-    return view_for_api_declaration
 
 
 class NodeWalker(object):
