+++ tests/acceptance/app/config.ini
@@ -5,6 +5,5 @@ pyramid_swagger.enable_response_validation = false
 pyramid_swagger.enable_path_validation = true
 pyramid_swagger.exclude_routes = /undefined/first
                                  /undefined/second
-pyramid_swagger.prefer_20_routes = /sample
 pyramid_swagger.schema_directory = tests/sample_schemas/good_app
-pyramid_swagger.swagger_versions = 1.2 2.0
+pyramid_swagger.swagger_versions = 2.0
