+++ tests/acceptance/config_ini_test.py
@@ -27,10 +27,9 @@ def test_load_ini_settings(ini_app):
     assert settings.validate_response is False
     assert settings.validate_path is True
     assert settings.exclude_routes == {'/undefined/first', '/undefined/second'}
-    assert settings.prefer_20_routes == {'/sample'}
 
 
 def test_get_swagger_versions(ini_app):
     settings = ini_app.app.registry.settings
     swagger_versions = get_swagger_versions(settings)
-    assert swagger_versions == {'1.2', '2.0'}
+    assert swagger_versions == {'2.0'}
