+++ tox.ini
@@ -1,15 +1,16 @@
 [tox]
-envlist = py27,py33,py34,py35,{py27,py35}-pyramid15
+envlist = py38,py310
 
 [testenv]
 skip_install = True
-deps = -rrequirements-dev.txt
-    pyramid15: pyramid<=1.5.4
+deps =
+    -rrequirements-dev.txt
+    pyramid
 
 commands =
     coverage run --source=pyramid_swagger/ --omit=pyramid_swagger/__about__.py -m pytest --capture=no --strict {posargs:tests/}
     coverage report -m
-    py27: pre-commit run --all-files
+    pre-commit run --all-files
 
 [flake8]
 exclude = .svn,CVS,.bzr,.hg,.git,__pycache__,.tox,*.egg,docs/conf.py
